requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
Werkzeug>=3.0.0